refgactoring of unit tests
